[FIX] account_followup: attach the actual invoice PDFs
Consider the case that we attach invoices to the followup.
We only attach invoices that have previously been printed / generated.
(Except if someone attaches a custom file which could be whatever.)

Currently:
  1. During the generation of the followup report we basically just
     regenerate the invoices by calling the invoice template.
  2. The PDFs that are added in the manual followup wizzard are not added to
     the generated PDF.

Problem: (Reproduce at the bottom)
Point (1.) is a problem in case the template does not reflect the actual
printed invoice. This can happen since the PDF output may also be altered during render time
(by overwriting '_render_qweb_pdf_prepare_streams').
This is i.e. a problem for l10n_ch. There, a QR code is appended
to the invoice PDF only during render time. This QR code does not
appear in the followup report since it is not part of the template.
Point (2.) may lead to missing invoices in the generated PDF.

After this commit:
The actual PDF (that was generated previously) is attached to the invoice.
Thus i.e. the QR code appears in l10n_ch.
In the process Point (2.) was also solved. In case the manually added
files are not PDFs they are silently dropped.

Reproduce
  1. Install l10n_ch
  2. Select 'CH Company'
  3. Create an Invoice
       * Due Date: in the past
       * Customer: Easy Clean Lausanne (A Swiss company)
  4. Print the PDF.
     Check that the QR code is appended to the invoice
  5. Go to follow-up reports for Easy Clean Lausanne
     and follow up with 'Print' option.
     Ensure 'Attach Invoices' is checked.
     Add some PDF file manually.
     Click "Print"
  6. The invoice is appended to the follow-up report
     but w/o QR code.
     The PDF that was manually added does not appear in the report at all.

task-3472962

closes odoo/enterprise#56451

X-original-commit: 5ecaa155224c678713df54c3db8598560c83229a
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Sven Führ (svfu) <svfu@odoo.com>