[IMP] documents: use an explicit thumbnail status to generate the preview
Purpose
=======
The thumbnails of the PDFs are generated by the browser of the first
user previewing the PDF. We will extend the building of the preview
and so we want something a bit more explicit than "False".

Rename hasThumbnail into hasStoredThumbnail to highlight the fact that
the thumbnail is stored in the database.

Task-3537562

Part-of: odoo/enterprise#49111