[IMP] documents: use the same route to preview the PDF and to preview other files
Purpose
=======
A custom route has been created to preview the PDF and check the write
access of the file. This was needed because the browser will generate
the preview, then make a write RPC call to change it.

Instead, we can use the same route for all files and make a route that
writes the thumbnail in SUDO, if we can read the record.

Task-3537562

Part-of: odoo/enterprise#49111