[FIX] web_studio: prevent editing the employee profile view
Steps to reproduce
==================

- Install hr,web_studio
- Login with admin
- Go to "My Profile" (top right)
- Open studio
- Add a field
- Close studio
- Login with demo
- Go to "My profile"

=> There is an access error

Cause of the issue
==================

On the profile view, there are some fields that are present
and the user has read access to using a whitelist SELF_READABLE_FIELDS.

If we add another field with studio, it will not be in the whitelist and
this will fail the permission check.

We cannot simply add a check to whitelist any manual fields, because
this could cause some security issues.

The res_users.get_view is also read called with sudo when the record is
matching the current user. This means groups applied to custom fields
from within studio won't restrict the field.

Solution
========

Disable the edition of that view.

opw-3598600

closes odoo/enterprise#54168

X-original-commit: de8fcb05801487da12ec4ba31fd0349928784233
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Hubert Van De Walle <huvw@odoo.com>