[FIX] l10n_cl_edi_pos: don't generate pdf for invoices 'boleta'
Issue:
For the Chilean localisation of POS, all orders generates the pdf of the invoices,
sadly this is a slow process (generating the pdf for 1 invoice takes around 4-5secs)

Cause:
A wkhtmltopdf bottleneck, since odoo/odoo@19916059c519c65083a05b270026420f8e41a062
we "Send and Print" the newly created invoices, which generates the pdf.
Some localizations have this as a requirement, but Chili apparently is not one of them.

Fix:
There are 2 types of invoices (boleta/factura), but the POS button "to_invoice" at order
checkout was used as a toggle between the types of invoices,
instead of a choice to invoice or not. And since the generation of the invoice is strongly coupled
with the generation of the pdf, I've added a context key to skip the pdf generation during POS checkout.
The user can still "Send & Print" the invoice from the backend.
The key is present when we *don't* select the invoice option at checkout (so for boleta invoices),
for factura the invoice's pdf is created in sync.

Affected versions:
17.0 up to master

Reference:
opw-3610333

closes odoo/enterprise#55175

X-original-commit: 22c629835c67779a9fa43a905d247e3b95d389f8
Related: odoo/odoo#151198
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>