[IMP] data_merge: hide technical fields from the chatter
When merging two records, all of the fields of the merged record are logged into the master record chatter.
We filter out some fields, such as file fields or fields with access rights,
but there are still many irrelevant technical fields and even some potentially sensitive fields (access tokens, etc.) being logged.

Moreover, with the introduction of the export_string_translation field flag,
those fields could not even be translated anymore.

The goal of this task is to exclude fields with the export_string_translation flag set to true
from the chatter logs to reduce clutter.

Task-3716421

closes odoo/enterprise#55602

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>