[REF] *: unify no content helper and sample data behavior
This commit corrects how the sample data styling is applied to the
cohort view by removing a duplicate rule and make it so that the styling
only concerns the table and not the view buttons. It also corrects how
the sample data styling rules are applied in the gantt view so that all
elements of the view are concerned by the rules, removing in the same
process the useless condition on disabled rows present in planning.

related to https://github.com/odoo/odoo/pull/148138

task-3650135

closes odoo/enterprise#53646

Related: odoo/odoo#148138
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>