[FIX] knowledge: resize name input on article change
This commit fixes a bug where the name input would not resize when switching
articles, leading to some titles being unreadable.

Now each time that the article changes we resize the input to accomodate the
new title.

Remove the erroneous `t-esc` usage.
Rectification of 7ea555fc76ad85c536a7d22495da8603dba064ce

Now that the topbar is a component, the span that is guaranteeing the width of
the container of the article name input can be set with a `t-out` instruction
(the topbar is re-rendered each time the record changes). The previously used
`t-esc` instruction was erroneous as it does not do anything in an owl template.

task-3640726

closes odoo/enterprise#53766

X-original-commit: 3d76686cbaedcaa7b9d552e2fc36a85bfa965d0e
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>