[FIX] knowledge: restore cursor position when closing /link dialog
When the user inputs the "/link" command and subsequently closes the
dialog for article selection, the cursor currently relocates to the
start of the document, disrupting the editing flow. To address this
issue, this commit will restore the cursor to its original position
upon dialog closure.

Steps to reproduce the issue:
1. Open an article you can edit
2. Put your cursor in a text block
3. Enter the "/link" command
4. Press `enter` to validate the command
5. Close the dialog for article selection (-> "CANCEL" button)

=> The cursor will be relocate to the start of the document.

TO BE: The cursor should be restored to its original position upon
dialog closure.

task-3620947

closes odoo/enterprise#53864

X-original-commit: 09af240b59c16264060765accbad6fde0102fc16
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>