[FIX] documents: fix incorrect file size displayed in the trash
How to reproduce:
- Open the app document
- Ensure the trash is empty
- Select a document from any workspace (not too small so that the displayed size
in MB is different from 0)
- Move it to the trash
- Open the trash
- Select the document and click on delete in the inspector
- Confirm the permanent deletion
The inspector displays 0 documents but the size is different from 0 MB.

This fixes the error by reloading the model forcing the re-computation of the
size.

Technical notes: no test has been added as the tour that test the deletion uses
small documents which prevents to add assertion about the size (as the size is 0
even when a document is present in the trash).

Task-3637482

closes odoo/enterprise#57599

X-original-commit: 9ebf25fa24a05a453365ae83e2f2744f63b9cfda
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>