[IMP] l10n_fr_reports: readability SIG report
In 16.3 and 16.4 we worked a lot on making the report more readable but the SIG
report was forgotten during those changes.
Commit:
- https://github.com/odoo/enterprise/commit/d2d4f5e04042eb420f0f4c566269d1148baf84f7
- https://github.com/odoo/odoo/commit/56a7f7745c5df9b70e6fe5b78f401085ce46284c
- https://github.com/odoo/enterprise/commit/538294661a5c0c748b55ad2d335448cd14700886

This commit will change the report to be more consistent with the other by
changing the disposition of the section and changing the hierarchy level either
by removing it and take advantage of the children_ids or changing to 0 to have
the css highlight.

closes odoo/enterprise#53895

Task: 3669714
X-original-commit: 76d9188767e26c4ae1f815334af5ca3ac6c5fcdf
Signed-off-by: William André (wan) <wan@odoo.com>