[FIX] l10n_co_edi: update for anexo 1.9
The spec for electronic invoices in Colombia was updated and is now
known as Anexo 1.9.
This commit updates the electronic invoice to meet the new  spec

A bit part of the new spec is the conversion of many fields from document to company currency / COP.

There is also a vendor document. But it follows a different spec.
Thus some shared logic / section with the electronic invoice was updated
so that the vendor document remains unchanged.

The following sections were removed since they are obsolete and also (basically) dead code:
  * OVT section: dead code
  * FE1 section: content is dead code

Previously when creating a credit note from an invoice with the
reversal wizard (account.move.reversal) there was the following bug.
The Credit Note Concept (l10n_co_edi_description_code_credit on account.move) was set
after posting. This was corrected in this commit (since a validation on post
was added to check that the concept is there).

task-3639271

closes odoo/enterprise#55279

Related: odoo/odoo#151431
Signed-off-by: Josse Colpaert <jco@odoo.com>