[IMP] approvals: UX improvements-multicompany
This commit adds the possibility for user to see the company
for which he wants to create an approval request if they have the
multi company access rights.
the company field is also added in the approval request tree view
for better visibility.

In addition to that, this commit a small ux improvement for the
approve button to make it more visible or less visible depending
if a document has been attached to the approval request or not.
For the tree view, the approvers are now displayed to be visible
for the user.

task-3632010

closes odoo/enterprise#53712

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>