[FIX] web_studio: report editor: save as an inheriting view
Before this commit, the report editor saved nodes (or rather, their *content*) as expressed
by their coordinates in the view they were originating from: (view_id, xpath), where
view_id is the actual view in db that has injected or modified that node. This is called
the branding (context key: "inherit_branding"), and is used by website for their edition.

Now, imagine a view that is inherited.
The inheriting view adds a node somewhere. All the resulting siblings of that node will be branded,
because under that philosophy, they represent an edition unit inside which it is safe to add/remove content. (see `ir_ui_view: def save`). The parent however loses its coordinates and cannot be considered
a safe place to add or remove content.
For our example, imagine we add a node in the main view, just below another one called "A".
If the inheriting view was taking node A as the xpath expression, then adding one before or after will
break the inheritance. See the message of 68bdd11b60cd75289795981d70e97697a38192c7 for a detailed example.

We can conclude that most use cases of report edition did not work, or at least that
limitations were dramatic and sometimes very dark.

This called for a change in modelisation, as a fix, otherwise the report editor
would be unusable for quite a long time for many people.

Thus, this commit offers a way out.
It proposes to save the whole arch of a view as an inheriting view, with a priority that allows
for it to be executed last, this is very important.
The arch of the inheriting view contains *all* the new arch for that view.

This is a bit heavy in DB, and for processing, as the inheritance mechanism will apply every
inheritance spec on top of the other as it should, and then, BOOM, the studio inheriting view comes
and every previous work has been useless because the whole thing is overriden anyway.

A more elegant solution may be worked out for master.

opw-3495170
opw-3568507
opw-3602609

closes odoo/enterprise#54429

X-original-commit: b7d0d3034d53cdab831f42f8b37a14cf50f3720e
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>