[IMP] knowledge: Improve HistoryDialog ui
Before this commit:
- When clicking on restore button, restoreRequested is called and the modal is closed.
In some cases, It's needed to do extra steps inside restoreRequested and the modal
should stay opened in background e.g: Adding a confirmation modal before restoring.
In this case (which was already done in related prs in project module), the history modal
is closed, so the user is not able anymore to see the revisions.

After this commit:
- Closing the modal was moved from _onRestoreRevisionClick to inside
restoreRequested to be executed according to the need. There is no impact in knowledge ux or ui.
It's just an adaptation as restoreRequested now request 2 params, so close was passed as a second
param.

Related Prs:
Community: https://github.com/odoo/odoo/pull/145459

task-3617834

closes odoo/enterprise#52351

Related: odoo/odoo#145459
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>