[ADD] l10n_br_{avatax,edi,edi_sale}_services: support services
This adds tax calculation and EDI support for service
transactions. All new functionality was added in new modules because
this is taking place in a stable release.

Avatax has different endpoints for submitting and cancelling service
invoices. These two endpoints have fairly similar payloads so various
overrides were made to remove/add information from the goods payloads
to make them compatible with the service endpoints. A new
account.external.tax.mixin l10n_br_is_service_transaction field is
used to determine whether to follow the old, existing goods flow or go
through the new service flow.

The biggest difference is that service invoices go through an extra,
asynchronous validation step. To achieve this, invoices that are
successfully submitted end up in a new "pending" state. After this,
their status can be checked with
button_l10n_br_edi_get_service_invoice(). This function is also called
periodically with a cron.

Service codes need to be configured per company, per product and per
city. Because of this we added a dependency on base_extended_address
and provided a list of cities [1]. A typical multi-company record rule
makes the service code visible in the right companies. This is the
first time we see a multi-company record used in external tax
calculation. To make this work in flows that can run as the public
user we set the right company in the two external tax portal
controller overrides before calculating taxes.

In master, these modules will be merged into the already existing
modules and deleted.

Thanks to MADI for the translations!

task-3660574

[1] From https://en.wikipedia.org/wiki/List_of_cities_in_Brazil_by_population

closes odoo/enterprise#61532

X-original-commit: 02e7e97f1c70be20d308cbaa44d8df809b823f0b
Related: odoo/odoo#163505
Signed-off-by: Josse Colpaert <jco@odoo.com>