[FIX] account_accountant: avoid annoying predictions
Avoid annoying the user when he put an account manually then changes the
label because the account changes again.
In order to do that, we only apply the predicted account if it is better
than the second option for prediction.

To reproduce:
* create an invoice
* add a line
* change the account
* change the label of the line

Previously, it would always change and override the account even if it
had a low certainty about the prediction.
Now, it will only override the current account if there is a prediction
that is clearly better.
The threshold is a ranking at least 10% better than the secondbest
choice.

task-3525714

closes odoo/enterprise#54537

X-original-commit: 20835179fd9b0f68b63ca292806df682abd190d8
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>