[IMP] website_sale_subscription: improved xpath expression
Since [1], this xpath was introduced. this xpath expression will generate
the issue in the following case:

For Instance, We have installed some custom module which is adding some
parameters in the `t-cache` attribute of view `product` of website_sale module.

After that, if I install the website_sale_subscription module then the product
page will be crashed. Because the xpath which is going to be added from the
module website_sale_subscription which is: `//t[@t-cache='pricelist,product']`
will not be able to find the newly added parameters by a custom module.

[1]: bc627f9

closes odoo/enterprise#76638

Closes: odoo#44086
X-original-commit: cbe062025991ee1c765a381de4f8532bab92d708
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>