[IMP] currency_rate_live: Use Bulgaria National Bank for exchange rates
Objective
---------
Add the Bulgarian National Bank (BNB) as a Currency Provider in Odoo.

Solution
---------
Rates are given against BGN in an XML file of format:

<ROWSET>
  <ROW>
    <GOLD>1</GOLD>
    <NAME_>Currency</NAME_>
    <CODE>Code</CODE>
    <RATIO>Per unit of currency/gold</RATIO>
    <REVERSERATE>Reverse rate for 1 BGN</REVERSERATE>
    <RATE>Levs (BGN)</RATE>
    <CURR_DATE>Date</CURR_DATE>
    <F_STAR>0</F_STAR>
 </ROW>
</ROWSET>

Source: https://www.bnb.bg/AboutUs/AUFAQ/Contr_Exchange_Rates_FAQ?toLang=_EN

Currencies returned on the 8th of January 2024:
    AUD, BRL, CAD, CHF, CNY, CZK, DKK, GBP, HKD, HUF, IDR, ILS, INR,
    ISK, JPY, KRW, MXN, MYR, NOK, NZD, PHP, PLN, RON, RUB*, SEK, SGD,
    THB, TRY, USD, ZAR, XAU
* Due to the situation on the international financial markets, the
Bulgarian National Bank (BNB) is not in a position to set a reference
exchange rate for the Bulgarian Lev against the Russian Rouble that is
representative of market conditions.

If a currency has no rate, it is skipped.

task-3672394

closes odoo/enterprise#54879

X-original-commit: 325c63b98b650cd5aecc73a4934d1849c571987b
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>