[FIX] timesheet_grid: allow to update timesheets of an archived employee
Before this commit, when the user tries to update a timesheet of an
archived employee, he got an error saying:
"Timesheets before the {last_validated_timesheet_date set on the employee}
(included) have been validated, and can no longer be modified"
The reason is because the employee is archived and so the check does not
fetch the employee to be sure the approver who alters the timesheet
is the manager/timesheet manager of the employee archived.

This commit fixes the issue by also fetching the archived employees
to be sure to check is correctly done as expected.

Versions impacted: saas~15.2+

opw-3613473

closes odoo/enterprise#54586

X-original-commit: fe0af8c8e6c485f2a2014fe781277255ac53fff2
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>