[FIX] delivery_ups_rest: send `CODAmount` as string
Steps to reproduce:
Set up a UPS shipping method, enable "Collect on Delivery" in the
configuration. Make a sale order with this shipping method, confirm
and validate the delivery.

Error:
```
Missing or invalid COD monetary value for Package  1. Valid length is 0 to 8 numeric and can hold up to 2 decimal places.
```

Based on UPS documentation, the value for COD needs to be of type
string.

opw-3666755

closes odoo/enterprise#54793

X-original-commit: 521090f40afdc58e2b720811c836198e1c475895
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>