[FIX] [test_]appointment[_hr,_full]: adapt query counters
Following: e02ec53a9e91343423b3f54a766bce4919631328

We introduced a new image_1920 field on the appointment.type model.

This field is read on the main appointment route, to show the appointment image
on the right side panel containing the meeting information.
Meaning we have an additional request towards the "attachment" model to fetch
that extra image.

We forgot to increment all query counters that hit that route by one.
This commit fixes it.

Task-3607606

closes odoo/enterprise#53912

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>