[FIX] knowledge: restore cursor position when closing calendar dialog
Similarly to this [commit], but for the `/calendar` command, when closing the
dialog, the article scrolls to the top.

This is because the focus is forcibly given back to the `editable` and the
selection was not put back inside at the position where the command was
executed: it is reset to the top instead.

This commit restores the selection when closing the dialog so that when the
focus is given back to the `editable`, no scrolling is done.

[commit]: https://github.com/odoo/enterprise/commit/09af240b59c16264060765accbad6fde0102fc16

task-3620947

closes odoo/enterprise#53933

X-original-commit: 2e3357c1cec4b17a8d1197c42d97694de882a032
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>