[IMP] whatsapp, test_whatsapp: improve template switch message
PURPOSE:

Currently while the conversation is going if a new template is shared
with the same customer then a message is dropped on the channel
which needs improvement in language construct.

SPECIFICATION:

Improve construct when document is sent by

a) same user using different model:
Post a message consisting of the template body in the channel.

b) Different user from any model:
"Chat shall be moved to DocumentName."

task-3607392

closes odoo/enterprise#54121

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>