[FIX] account_avatax: skip address validation for public user
Currently, a public user can't add any product to the cart when using
GeoIP and Avatax.

Steps to reproduce
------------------
* setup GeoIP
* setup Avatax credentials
* enable "Detect Automatically" on the "Automatic Tax Mapping (AvaTax)"
  fiscal position
* access to the website as a public user with an IP address from the US
* try adding a product to the cart.

You should be met with a validation pop

Cause
-----
For the Avatax fiscal position to work, we need the partner's country,
state and zip code. Usually, this isn't a problem because the fiscal
position is set after the Public User enters their address. But, when
using GeoIp, the fiscal position is set right when the user lands on the
page, based on their location data. This triggers a check for the
address, but the Public User hasn't entered one yet, which causes a
validation error.

opw-3625410

closes odoo/enterprise#55195

X-original-commit: 1820955d24bc9bd0f5b964b64f59f42dab80010e
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Séna Serge Nshimiyimana (sesn) <sesn@odoo.com>