[IMP] web_studio: add open_form_view attribute support
This commit adds the open_form_view attribute support in the sidebar of
the List editor. The option is only visible when edition is enabled on the
view (top or bottom).

When enabled, the 'View' buttons appear directly inside Studio. This required
to edit the template of the list renderer, since the buttons are only displayed
when an associated method to open the form view is present (which is not the
case here).

A test has also been added to the list view editor tests.

task-3605574

closes odoo/enterprise#54022

Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>