[FIX] studio: prevent field fetching in subviews when studio_no_fetch
In Studio, creating a many2many field towards a model containing a
group restricted field causes a crash.

The group restricted field in the arch is tagged with the
`studio_no_fetch` attribute to avoid studio from fetching it. However, this
attribute is not parsed when placed in subviews.

This commit prevents the group restricted field in subviews
from being fetched by studio.

Steps to reproduce:

- Install CRM, Studio and Project
- Go to CRM, open a record
- Open Studio
- Create a new Many2many field with a relation to Project
- Crash

opw-3603635

closes odoo/enterprise#54015

Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>