[FIX] knowledge: disable magic selection when clicking on the title field
In Knowledge, the system automatically selects the entire text of the
title field when the user clicks on the field. When the user clicks on
the field again to reposition the cursor or selects a portion of the
text, the system will re-select the entire text which prevent people
from easily editing the title field.

Users are required to utilize the left or right arrow keys on their
keyboard to move their cursor or employ keyboard shortcuts to select a
specific portion of the text which is not convenient.

To enable text selection, we will disable the magic selection. The
title field will now behave like any other text field of Odoo.

task-3676051

closes odoo/enterprise#54306

X-original-commit: aa3c5e5de52e78b33b873226c172c61460b7d2cc
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>