[IMP] knowledge: add /clipboard demonstration video in demo article
Currently, users may not know that the /clipboard command exists, may not
know what they can do with it and may not know how to use it. To improve
the onboarding experience and showcase the /clipboard command, we will
integrate a demonstration video in the demo article.

After watching this demonstration video, people should better understand
the purpose of the /clipboard command and should know how to use it.

task-3676051

closes odoo/enterprise#54122

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>