[IMP] knowledge: add AI generation prompt for empty articles
Now that Odoo is using AI prompts for the edition of articles and website,
we are adding this feature to the complete generation of articles.

This commit adds a new button when article is empty besides the other already
present.
This button opens a dialog which contains the prompt to interact with ChatGPT,
this way the user gets the possibility to get generated articles and can insert
the one they prefer.

task-3619294

closes odoo/enterprise#54179

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>