[FIX] web_enterprise: traceback when unselecting all measures on mobile
Before this commit user was getting traceback on unselecting
all measures in pivot view on mobile.

Steps to reproduce:
- ​switch to mobile view.
- open the pivot view of any report (tested with the 'tasks analysis' one here).
- unselect all measures, the traceback will appear.

Observed behavior:
a traceback is generated.

Expected behavior:
no traceback, the action helper is displayed

After this commit user will not get traceback on unselecting
all measures in pivot view on mobile. Instead of that an
action helper will be displayed.

Task-3630703

closes odoo/enterprise#57267

X-original-commit: 4a273f19677e709d3c2f8b76c1d6ca4e8c9ea23a
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>