[FIX] l10n_co_edi: add a test for the vendor document
Previously there was no test for the vendor document.
This commit adds a (simple) test.c

In the next commit of the same PR the code for the vendor document had to
be adapted while keeping the result the same.
This test helps in making sure nothing went wrong.
The reason was that the electronic invoice and vendor document share
some common code / logic but the shared parts had to be changed for an
update to the electronic invoice spec (Anexo 1.9).
(See the next commit for details)

task-3639271

X-original-commit: 7586a92bb577349c920a7ba7827a9e693558860d
Part-of: odoo/enterprise#61531