[IMP] social_facebook: show the users reactions to the posts / comments
Purpose
=======
Currently, we only show the likes of the users. But they can also
react with something else than a like. This commit shows the emojis
of the 2 most used reactions (only 2, because we don't have that
many space on the card) for the post, for the comments and replies
we show all emojis, because we have more space.

Remove unused import.

Task-3595484

Part-of: odoo/enterprise#54187