[FIX] hr_appraisal: correct image mismatch
Steps to reproduce the bug:
1. Have a user that have a user image that is different from the employee image.
2. Go to an appraisal form of that employee. (or create one) -> the miniature is different from the full image.

expected result:
the miniature should be the same as the full image if possible.

correction:
The image of the right as been removed because browsers will cache the
small one even if not stored in db and will not update it when The
employee image is updated. So in some cases the problem is just not
solvable without clearing the browser data even if all is done properly.

task-3644834

closes odoo/enterprise#57733

X-original-commit: 84a627f06450dd8e595e6c381e07343888ae1c57
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>