[FIX] knowledge: unset max-width constraint on form sheet
Currently, there is a CSS rule constraining all form sheets of Odoo to
have a maximal width of 2600px. Since the Knowledge editor is a form
view, it is affected by this rule. As a result, the editor does not
fully utilize the available space on large displays, leaving blank
space on the right side of the screen.

To enhance user comfort, we unset the max-width constraint of the form
sheet of Knowledge. The page now takes all available space and the
editor is now centered.

task-3636625

closes odoo/enterprise#54565

X-original-commit: d191d25f6df8ac50dcd8ee7368f04158ac1923dc
Signed-off-by: Julien Banken (jbn) <jbn@odoo.com>
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>