[FIX] account_accountant: reconcile foreign currencies with custom rate
Suppose the following reconciliation situation:
Company's currency = EUR
We want to reconcile
105 USD = 98 EUR
with
105 CAD = 73 EUR

Before this PR:
When encountering two foreign currencies Odoo will try to reconcile
in EUR (company's currency).
After the first step of reconciliation there is a residual of 25 EUR = 27 USD
left on the first line. We therefore try to create a write-off of 25 EUR
and reconcile it with the residual, those 25 EUR are converted in USD at
the Odoo rate => 24 USD.
27 USD != 24 USD, the reconciliation can't be full while it should have.

Now we detect cases where there is only one residual left in multicurrency and
create the write-off in that currency.

opw-3497793

closes odoo/enterprise#57176

X-original-commit: e84cbd01afa7703a1ec630ee99051bea645f03e0
Signed-off-by: Laurent Smet (las) <las@odoo.com>
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>