[FIX] account_reports: properly compute default external values on tax reports when posting the tax closing
Before this fix, when the tax closing was posted, though the tax lock date was properly updated, the default external values (computed from _default expressions) of the tax report were not computed. It only worked when setting the tax lock date manually from the lock wizard.

We take advantage of this commit to improve the test suite in order to check all possible cases for the creation of default external values.

closes odoo/enterprise#55001

X-original-commit: fb57765de65839336530c0f9204f2d64a62a8aa7
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: John Laterre (jol) <jol@odoo.com>