[IMP] knowledge: improve the TOC behavior's usability
This commit updates the table of content behavior so that it is easier to
understand how to use it for users.
It lets the string present inside this behavior be visible when one or less
heading is found so that the users understand that they need to add some to
fill the table of content.

We also removed the class `oe_unremovable` from the links inside the TOC
because they are no longer necessary because of two attributes:
  * contenteditable='false'
  * data-oe-protected='true'.
Plus they created issues with the insertion of the behavior inside the editable.

task-3676903

closes odoo/enterprise#54099

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>