[FIX] documents: handle delete button visibility
Before this commit:
The Delete button was visible in all instances of the document form view,
including documents, projects, etc even while creating a new record. This was
not logical since the Delete button is only useful when editing a workspace,
not while creating a new record.

Technical:
We have added an `attrs` having invisible with the condition
`('id', '=',  'False')` to the delete button, so It won't be visible if the
record is not created.

After this commit:
The Delete button will only be visible in the Documents app when editing a
workspace.

Task-3607426

closes odoo/enterprise#54071

X-original-commit: c5988b9766d471b7c8730de88234cf404533780b
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>