[FIX] social_facebook: show the stickers in the comment modal
Purpose
=======
Currently, if someone sends a sticker in a Facebook comment, a blank
bubble is displayed in the feed view. Instead, we want to display the
sticker.

From the API pov, stickers are just images.

If there's only an image without text, show the message without the blue bubble.

Task-3595484

closes odoo/enterprise#61169

X-original-commit: d8c04422e741048779fbedc105d75d02c4f9268b
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>